pyproject.toml
[tool.poetry]
name = "django-gallery"
version = "0.1.0"
description = ""
authors = ["Michał Bońkowski <michal.bonkowski@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1"
djangorestframework = "^3.15.2"
django-stubs = {extras = ["compatible-mypy"], version = "^5.0.4"}
ruff = "^0.6.3"
pytest-django = "^4.9.0"
celery = "^5.4.0"
gunicorn = "^23.0.0"
pillow = "^10.4.0"
requests = "^2.32.3"
redis = "^5.0.8"
types-requests = "^2.32.0.20240712"
pytest-mock = "^3.14.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
mypy = "^1.11.2"
pylint-to-ruff = "^0.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
